feat: Improve the douki structure to allow support for more programming languages
