+++ pyproject.toml
@@ -30,6 +30,7 @@ packages = [
 include = [
   "src/douki/py.typed",
   "src/douki/schema.json",
+  "src/douki/_base/schema.json",
 ]
 exclude = [
   ".git/*",
@@ -114,7 +115,7 @@ quote-style = "single"
 [tool.vulture]
 exclude = ["tests"]
 ignore_decorators = []
-ignore_names = []
+ignore_names = ["cwd", "paths", "excludes"]
 make_whitelist = true
 min_confidence = 80
 paths = ["./"]
