+++ src/douki/_base/__init__.py
@@ -1,0 +1,3 @@
+"""
+title: Base (language-agnostic) components for Douki.
+"""
