+++ src/douki/_base/schema.json
@@ -1,0 +1,162 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://example.com/douki/schema.json",
+  "title": "douki YAML docstring",
+  "type": "object",
+  "additionalProperties": false,
+
+  "definitions": {
+    "param_entry": {
+      "anyOf": [
+        { "type": "string" },
+        {
+          "type": "object",
+          "additionalProperties": false,
+          "properties": {
+            "type": { "type": "string" },
+            "optional": { "type": ["boolean", "null"] },
+            "description": { "type": "string" },
+            "default": {},
+            "variadic": {
+              "type": "string",
+              "enum": ["positional", "keyword"]
+            }
+          }
+        }
+      ]
+    },
+    "typed_entry": {
+      "type": "object",
+      "additionalProperties": false,
+      "properties": {
+        "type": { "type": "string" },
+        "description": { "type": "string" }
+      }
+    },
+    "example_entry": {
+      "anyOf": [
+        { "type": "string" },
+        {
+          "type": "object",
+          "additionalProperties": false,
+          "properties": {
+            "code": { "type": "string" },
+            "description": { "type": "string" }
+          }
+        }
+      ]
+    }
+  },
+
+  "properties": {
+    "title": { "type": "string" },
+    "summary": { "type": ["string", "null"] },
+    "deprecated": { "type": "string" },
+
+    "visibility": {
+      "type": "string",
+      "enum": ["public", "private", "protected", "internal"]
+    },
+    "mutability": {
+      "type": "string",
+      "enum": ["mutable", "immutable", "constant"]
+    },
+    "scope": {
+      "type": "string",
+      "enum": ["instance", "static", "class"]
+    },
+
+    "parameters": {
+      "type": "object",
+      "additionalProperties": { "$ref": "#/definitions/param_entry" }
+    },
+
+    "returns": {
+      "anyOf": [
+        { "type": "string" },
+        { "type": ["string", "null"] },
+        { "$ref": "#/definitions/typed_entry" }
+      ]
+    },
+    "yields": {
+      "anyOf": [{ "type": "string" }, { "$ref": "#/definitions/typed_entry" }]
+    },
+    "receives": {
+      "anyOf": [{ "type": "string" }, { "$ref": "#/definitions/typed_entry" }]
+    },
+
+    "raises": {
+      "anyOf": [
+        {
+          "type": "object",
+          "additionalProperties": { "type": "string" }
+        },
+        {
+          "type": "array",
+          "items": { "$ref": "#/definitions/typed_entry" }
+        }
+      ]
+    },
+
+    "warnings": {
+      "anyOf": [
+        {
+          "type": "object",
+          "additionalProperties": { "type": "string" }
+        },
+        {
+          "type": "array",
+          "items": { "$ref": "#/definitions/typed_entry" }
+        }
+      ]
+    },
+
+    "see_also": {
+      "anyOf": [
+        { "type": "string" },
+        {
+          "type": "array",
+          "items": { "type": "string" }
+        }
+      ]
+    },
+
+    "notes": { "type": ["string", "null"] },
+    "references": {
+      "anyOf": [
+        { "type": ["string", "null"] },
+        {
+          "type": "array",
+          "items": { "type": "string" }
+        }
+      ]
+    },
+
+    "examples": {
+      "anyOf": [
+        { "type": "string" },
+        {
+          "type": "array",
+          "items": { "$ref": "#/definitions/example_entry" }
+        }
+      ]
+    },
+
+    "attributes": {
+      "type": "object",
+      "additionalProperties": { "$ref": "#/definitions/param_entry" }
+    },
+
+    "methods": {
+      "anyOf": [
+        { "type": "string" },
+        {
+          "type": "array",
+          "items": { "type": "string" }
+        }
+      ]
+    }
+  },
+
+  "required": ["title"]
+}
