+++ src/douki/_python/__init__.py
@@ -1,0 +1,3 @@
+"""
+title: Python language backend for Douki.
+"""
